Introduce BaseCollector class
We decided to provide a base class for all collectors to provide methods
and properties which are useful and/or needed by other collectors.
First property is here `controller_name_from_class` which calculate the
collector name from `__class__.__name__`.
As we want to use `CollectorBase` to provide common features to all
collectors we moved loading of collector options from config to it.
Within this Change we also rename the example collector from `my` to
`example` to be more common in naming.