Remove p3.yml and deliver an example instead
As we want to let work the code on each system for development we
removed the p3.yml from source and put it on gitignore.

Now we manage an example of the configuration file which has to be
copied and adapt for the local development.

While docker build we move the example file to the real `p3.yml` as the
example should ever work with linux systems.