Fix container build
There was an error in building container immage as python expect to use
venv instead of installing p3exporter to system python. As this
container is only for illustrating how it can work we decided to simply
add optione `--break-system-packages` to pip command to override the
default behavior.

We also moved the container manifest to the more general `Containerfile`.